requirements.txt
pycryptodome-test-vectors==1.0.8
pycryptodomex==3.14.1
